[Foundation] Implemented mkdir
