+++ Tests/Functional/ModuleMapTests.swift
@@ -8,7 +8,7 @@
  See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 */
 
-import func POSIX.mkdir
+import func Utility.mkdir
 import func POSIX.popen
 import Utility
 import XCTest
