+++ Tests/Functional/ValidLayoutTests.swift
@@ -12,7 +12,7 @@ import struct Utility.Path
 import func Utility.symlink
 import func Utility.walk
 import func POSIX.rename
-import func POSIX.mkdir
+import func Utility.mkdir
 import func POSIX.popen
 import XCTest
